refactor(texture): split texture loading responsibilities
Extract DDS validation, GL upload, texture loading, fallback/cache ownership, and PreviewTexture into focused internal modules.

Keep TextureManager as the public facade and add a checked-in gli swizzle translation patch for clang-tidy compatibility.
