+++ src/TextureLoader.cpp
@@ -1,0 +1,178 @@
+#include "TextureLoader.h"
+#include "ArchiveAccess.h"
+#include "DdsTextures.h"
+#include "MoDataPaths.h"
+#include "PreviewNif.h"
+#include "PreviewTexture.h"
+#include "TextureUpload.h"
+
+#include <libbsarch/bs_archive.h>
+
+#include <QDebug>
+#include <QDir>
+#include <QFileInfo>
+
+#include <cstddef>
+#include <exception>
+#include <utility>
+
+TextureLoader::TextureLoader(MOBase::IOrganizer* organizer, TextureSourceProvider textureSource)
+    : m_MOInfo {organizer}
+    , m_TextureSource {std::move(textureSource)} {}
+
+std::unique_ptr<PreviewTexture> TextureLoader::load(const QString& texturePath) const {
+    if (auto texture = tryLoadFromSource(texturePath)) {
+        return texture;
+    }
+
+    return loadAuto(texturePath);
+}
+
+std::unique_ptr<PreviewTexture> TextureLoader::loadAuto(const QString& texturePath) const {
+    if (texturePath.isEmpty()) {
+        return nullptr;
+    }
+
+    if (!m_MOInfo) {
+        qCritical("Failed to interface with Mod Organizer");
+        return nullptr;
+    }
+
+    const auto* const game = m_MOInfo->managedGame();
+    if (!game) {
+        qCritical("Failed to interface with managed game plugin");
+        return nullptr;
+    }
+
+    const auto realPath = MoDataPaths::resolveDataPath(m_MOInfo, texturePath);
+    const bool fileExists = !realPath.isEmpty() && QFileInfo::exists(realPath) && QFileInfo(realPath).isFile();
+
+    if (fileExists) {
+        return loadLooseTexture(realPath);
+    }
+
+    if (auto texture = tryLoadFromMods(texturePath)) {
+        return texture;
+    }
+
+    if (auto texture = tryLoadFromGame(texturePath)) {
+        return texture;
+    }
+
+    return nullptr;
+}
+
+std::unique_ptr<PreviewTexture> TextureLoader::tryLoadFromSource(const QString& texturePath) const {
+    if (m_TextureSource.kind
+        == TextureSourceProviderKind::Auto
+        || !textureProviderCoversPath(m_TextureSource, texturePath)) {
+        return nullptr;
+    }
+
+    switch (m_TextureSource.kind) {
+        case TextureSourceProviderKind::Mod:
+        case TextureSourceProviderKind::GameData: {
+            if (!m_TextureSource.sourcePath.isEmpty()) {
+                const auto realPath = QDir(m_TextureSource.sourcePath)
+                                          .absoluteFilePath(QDir::cleanPath(normalizeTextureDataPath(texturePath)));
+                if (QFileInfo::exists(realPath) && QFileInfo(realPath).isFile()) {
+                    if (auto texture = loadLooseTexture(realPath)) {
+                        return texture;
+                    }
+                }
+            }
+
+            return tryLoadFromArchives(m_TextureSource.archivePaths, texturePath);
+        }
+        case TextureSourceProviderKind::Auto: return nullptr;
+    }
+
+    return nullptr;
+}
+
+std::unique_ptr<PreviewTexture> TextureLoader::loadLooseTexture(const QString& path) {
+    try {
+        auto texture = DdsTextures::loadFileIfValid(path);
+        if (texture.empty()) {
+            qWarning("Failed to decode loose DDS '%s': invalid or unsupported DDS", qUtf8Printable(path));
+            return nullptr;
+        }
+        return TextureUpload::upload(texture);
+    } catch (const std::exception& e) {
+        qWarning("Failed to decode loose DDS '%s': %s", qUtf8Printable(path), e.what());
+        return nullptr;
+    }
+}
+
+std::unique_ptr<PreviewTexture> TextureLoader::tryLoadFromArchives(
+    const QStringList& archivePaths,
+    const QString& texturePath
+) {
+    for (const auto& archivePath : archivePaths) {
+        if (auto texture = loadFromArchive(archivePath, texturePath)) {
+            return texture;
+        }
+    }
+
+    return nullptr;
+}
+
+std::unique_ptr<PreviewTexture> TextureLoader::tryLoadFromMods(const QString& texturePath) const {
+    if (!m_MOInfo) {
+        return nullptr;
+    }
+
+    const auto fileOrigins = m_MOInfo->getFileOrigins(texturePath);
+    if (fileOrigins.empty()) {
+        return nullptr;
+    }
+
+    const auto& modName = fileOrigins.constFirst();
+    if (auto* const mod = m_MOInfo->modList()->getMod(modName)) {
+        if (auto texture = tryLoadFromArchives(MoDataPaths::archivePathsFromMod(mod), texturePath)) {
+            return texture;
+        }
+    }
+    return nullptr;
+}
+
+std::unique_ptr<PreviewTexture> TextureLoader::tryLoadFromGame(const QString& texturePath) const {
+    if (!m_MOInfo) {
+        return nullptr;
+    }
+
+    return tryLoadFromArchives(MoDataPaths::archivePathsFromGame(m_MOInfo), texturePath);
+}
+
+std::unique_ptr<PreviewTexture> TextureLoader::loadFromArchive(const QString& archivePath, const QString& texturePath) {
+    libbsarch::bs_archive archive;
+    if (!ArchiveAccess::loadArchive(archive, archivePath)) {
+        return nullptr;
+    }
+
+    const auto buffer = ArchiveAccess::extractBytes(archive, texturePath);
+    if (buffer.isEmpty()) {
+        return nullptr;
+    }
+
+    try {
+        auto texture = DdsTextures::loadIfValid(buffer.constData(), static_cast<std::size_t>(buffer.size()));
+        if (texture.empty()) {
+            qWarning(
+                "Failed to decode BSA DDS '%s' from '%s': invalid or unsupported DDS",
+                qUtf8Printable(texturePath),
+                qUtf8Printable(archivePath)
+            );
+            return nullptr;
+        }
+        return TextureUpload::upload(texture);
+    } catch (const std::exception& e) {
+        qWarning(
+            "Failed to decode BSA DDS '%s' from '%s': %s",
+            qUtf8Printable(texturePath),
+            qUtf8Printable(archivePath),
+            e.what()
+        );
+        return nullptr;
+    }
+}
